修改平均出块时长统计方式

+++ pojo/src/main/java/com/platon/browser/dto/StatisticInfo.java
@@ -10,7 +10,7 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 @Data
 public class StatisticInfo {
-    private long avgTime;
+    private BigDecimal avgTime;
 
     private long current;
